requirements.txt
line-bot-sdk==3.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0
flask==2.3.3
gunicorn==21.2.0
